STY: flake8 + rearranging
